// frontend/src/backend.d.ts
export interface ShopDetails {
    address: {
        street: string;
        country: string;
        city: string;
        zipcode: string;
    };
    openingHours: {
        tuesday?: string;
        wednesday?: string;
        saturday?: string;
        thursday?: string;
        sunday?: string;
        friday?: string;
        monday?: string;
    };
    shopName: string;
    companyDetails: {
        taxId: string;
        vatId: string;
    };
    contactDetails: {
        email: string;
        phone: string;
    };
}